move getAllDatesInRange & tests to TrendsContainer, standardize timePrefs param
